fix: fix a bug for HyHex2Int2Str func

+++ src/hy_utils.c
@@ -135,7 +135,8 @@ void HyHex2Int2Str(char *str, uint8_t str_len, char *addr, uint8_t addr_len)
 {
     uint8_t len = 0;
     for (int i = 0; i < addr_len; i++) {
-        len += snprintf(str + len, str_len - len, "%03d", addr[i]);
+        // LOGD("addr: %02x, %d \n", addr[i], (uint8_t)addr[i]);
+        len += snprintf(str + len, str_len - len, "%03d", (uint8_t)addr[i]);
     }
 }
